pyproject.toml
[tool.poetry]
name = "social-media-downloader"
version = "1.0.0"
description = "Download content from Xiaohongshu and Weibo - images, text, and comments"
authors = ["Your Name"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
playwright = "^1.40.0"
httpx = "^0.27.0"
typer = ">=0.12.0"
rich = "^13.7.0"
pydantic = "^2.5.0"
beautifulsoup4 = "^4.12.0"
lxml = "^5.1.0"
pillow = "^12.1.1"

[tool.poetry.group.dev.dependencies]
# pyinstaller requires Python <3.14, install separately if needed:
# pip install pyinstaller

[tool.poetry.scripts]
smd = "src.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
